+++ cmd/deviceshifu/cmdOPCUA/main.go
@@ -4,8 +4,8 @@ import (
 	"log"
 	"os"
 
-	deviceshifuopcua "github.com/edgenesis/shifu/pkg/deviceshifu/deviceshifuOPCUA"
 	"github.com/edgenesis/shifu/pkg/deviceshifu/deviceshifubase"
+	"github.com/edgenesis/shifu/pkg/deviceshifu/deviceshifuopcua"
 
 	"k8s.io/apimachinery/pkg/util/wait"
 )
