+++ pkg/deviceshifu/deviceshifumqtt/deviceshifumqttconfig.go
