+++ pkg/deviceshifu/deviceshifumqtt/deviceshifumqttconfig_test.go
